requirements.txt
numpy==2.4.2
opencv-python==4.13.0.92
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.18.0
pyserial==3.5
PyYAML==6.0.3
simpleitk==2.5.3


# pip install -r requirements.txt
# python -m pip freeze | Out-File -Encoding utf8 requirements.txt